+++ tests/Mockery/ExpectationTest.php
@@ -240,6 +240,17 @@ class ExpectationTest extends MockeryTestCase
     }
 
     /**
+     * @test
+     * @requires PHP 7.0.0
+     */
+    public function it_can_throw_a_throwable()
+    {
+        $this->setExpectedException(\Error::class);
+        $this->mock->shouldReceive('foo')->andThrow(new \Error());
+        $this->mock->foo();
+    }
+
+    /**
      * @expectedException OutOfBoundsException
      */
     public function testThrowsExceptionBasedOnArgs()